Move ProvidedAdditionalChelcks & ConfigurableJwtValidator in their own files
